refactor(server): Split per-feature dependencies using modules
